Datos adicionales 3
- Se completa la generación de la tabla de datos adicionales al hacer click en los nodos apropiados. Falta darle formato con CSS.

- Los datos adicionales se guardan en archivos .json que tienen como título el objeto del estudio (es decir, el nombre del artista o la canción).

- Se cambian las rutas para los archivos .json y para cleanDataset.csv.

- Se añade cleanDataset.csv (Se había perdido en algún momento).
