+++ app/model/unifyDataset.py
@@ -1,4 +1,5 @@
 import pandas as pd
+import os
 
 d1 = pd.read_csv('check1.csv')
 d2 = pd.read_csv('check1.csv')
@@ -10,4 +11,4 @@ data = pd.concat(data,sort=False)
 
 df = data.drop(data[data.Song == '0'].index)
 
-df.to_csv('cleanDataset.csv',index=False)
+df.to_csv(os.getcwd() + '/app/static/cleanDataset.csv',index=False)
