+++ app/static/js/graph.js
@@ -200,43 +200,3 @@ network.moveTo({
 network2.moveTo({
   scale: 0.5             // Zooms out
 });
-
-// WIP
-network.on("click", function(params) {
-  params.event = "[original event]";
-  if (params.nodes.length > 0) {
-
-    var name = params.nodes[0];
-
-
-    $.getJSON('/app/static/songData1.json')
-        .done(function (data) {
-
-            var len = Object.keys(data.ID).length;
-            var table = document.getElementById("mytable");
-
-            addRow(table, data.ID[0], "");
-
-            for (i = 0; i < len; i++) {
-                var lastRowCells = table.rows[table.rows.length-1].cells;
-
-                if (data.idPropertyName[i] == lastRowCells[0].innerHTML) {
-                    lastRowCells[1].innerHTML += ", " + data.valueProperty[i];
-                }
-                else {
-                    addRow(table, data.idPropertyName[i], data.valueProperty[i]);
-                }
-            }
-
-        })
-  }
-});
-
-
-function addRow(table, data1, data2) {
-    var row = table.insertRow(-1);
-    var cell1 = row.insertCell(0);
-    var cell2 = row.insertCell(1);
-    cell1.innerHTML = data1;
-    cell2.innerHTML = data2;
-};
