+++ app/view/graph.html
@@ -63,7 +63,7 @@
 
     </div> <!-- Row -->
 
-    <pre id="eventSpan"></pre>
+    <table id="mytable"></table>
 
   </div> <!-- Container -->
